reorganize components and index by adding Main.js
