+++ client/components/index.js
@@ -1,5 +1,7 @@
-export {default as Cursor} from './Cursor'
-export {default as Floor} from './Floor'
+export { default as Cursor } from './Cursor'
+export { default as Floor } from './Floor'
 export { default as FirstVendor } from './FirstVendor'
 export { default as FirstVendorStoreFront } from './FirstVendorStoreFront'
-
+export { default as Player } from './Player'
+export { default as Main } from './Main'
+export { default as HomeScreen } from './HomeScreen'
